\name{mixtox}
\alias{mixtox}
\docType{package}
\title{
Dose-Response Curve Fitting and Mixture Toxicity Assessment
}
\description{Curve Fitting for monotonic(sigmoidal) & non-monotonic(J-shaped) 
concentration-response data. Prediction of mixture toxicity based on reference 
models such as 'concentration addition', 'independent action', and 'generalized 
concentration addition'.
}
\details{
\tabular{ll}{
Package: \tab mixtox\cr
Type: \tab Package\cr
Version: \tab 1.4.0\cr
Date: \tab 2022-06-19\cr
License: \tab GPL-2\cr
}
(1) Curve fitting of concentration-responses based on 13 monotonic(sigmoidal) and 4 
non-monotonic(J-shaped) models, goodness of fit, and calculation of confidence interval 
. \cr
(2) Experimental design for mixture toxicity.  
	acr: arbitrary concentration ratio; 
	eecr: equal effect concentration ratio;
	udcr: uniform design concentration ratio. \cr
(3) Mixture toxicity prediction based on reference models such as concentration addition (CA), 
	independent action (IA), and generalized concentration addition (GCA).
}
\author{
xiangwei zhu \cr
Maintainer: xiangwei zhu <xwzhunc@gmail.com>
}
\references{
Zhu, X.-W., Chen, J.-Y. (2016). mixtox: An R Package for Mixture Toxicity Assessment. 
R Journal, 8(2), 422-434.\cr
Zhu, X.-W., Ge, H.-L., Cao, Y.-B. (2016). Mixture cytotoxicity assessment of ionic liquids 
and heavy metals in MCF-7 cells using mixtox. Chemosphere, 163, 544-551.\cr
Zhu X-W, Liu S-S, Qin L-T, Chen F, Liu H-L. 2013. Modeling non-monotonic dose-response 
relationships: Model evaluation and hormetic quantities exploration. Ecotoxicology and 
Environmental Safety 89:130-136.\cr
Dunnett, C. W. 1964. New Tables for Multiple Comparisons with a Control. Biometrics 
30(3):482-491.\cr
Hickernell, Fred J. 1996. A Generalized Discrepancy and Quadrature Error Bound. Mathematics of 
Computation 67(211):299-322.\cr
Howard, Gregory J., Jennifer J. Schlezinger, Mark E. Hahn, and Thomas F. Webster. 2010. 
Generalized Concentration Addition Predicts Joint Effects of Aryl Hydrocarbon 
Receptor Agonists with Partial Agonists and Competitive Antagonists.  
Environmental Health Perspectives 118(5):666-672.\cr
Scholze, M. et al. 2001. A General Best-Fit Method for Concentration-Response Curves and the 
Estimation of Low-Effect Concentrations. Environmental Toxicology and Chemistry 
20(2):448-457.\cr
Wang, Yuan and Kai-Tai Fang. 1996. Uniform Design of Experiments with Mixtures. Science in China 
Series A-Mathematics Physics Astronomy 39(3):264-275.\cr
Backhaus, T., Faust, M., 2012. Predictive environmental risk assessment of chemical mixtures: A 
conceptual framework. Environmental Science and Technology. 46, 2564-2573.\cr

}
\keyword{mixture toxicity}
\keyword{goodness of fit}
\keyword{sigmoidal dose-response curve}
\keyword{non-monotonic dose-response curve}
\keyword{curve fitting}
\keyword{concentration addition}
\keyword{independent action}
\keyword{generalized concentration addition}
