\name{mixtox-package}
\alias{mixtox-package}
\alias{mixtox}
\docType{package}
\title{
Curve Fitting and Mixture Toxicity Assessment
}
\description{Curve Fitting for monotonic(sigmoidal) & non-monotonic(J-shaped) 
concentration-response data. Toxicity prediction of chemical mixtures using reference 
models such as 'concentration addition' and 'independent action' (Backhaus, T., Faust, 
M., Environmental Science and Technology. 2012, 46, 2564-2573. doi:10.1021/es2034125).
}
\details{
\tabular{ll}{
Package: \tab mixtox\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2015-11-30\cr
License: \tab GPL-2\cr
}
(1) Curve fitting of concentration-response data using 13 monotonic(sigmoidal) and 6 
non-monotonic(J-shaped) equations, statistics for goodness of fit , and calculation for 
function-based and observation-based confidence intervals . \cr
(2) Experimental design for mixture toxicity.  
	acr: arbitrary concentration ratio; eecr: equal effect concentration ratio; 
	udcr: uniform design concentration ratio. \cr
(3) Mixture toxicity prediction using reference models such as concentration addition (CA), 
	independent action (IA), and generalized concentration addition (GCA).
}
\author{
xiangwei zhu \cr
Maintainer: xiangwei zhu <xwzhunc@gmail.com>
}
\references{
Dunnett, C. W. 1964. New Tables for Multiple Comparisons with a Control. Biometrics 
30(3):482-491.\cr
Hickernell, Fred J. 1996. A Generalized Discrepancy and Quadrature Error Bound. Mathematics of 
Computation 67(211):299-322.\cr
Howard, Gregory J., Jennifer J. Schlezinger, Mark E. Hahn, and Thomas F. Webster. 2010. 
Generalized Concentration Addition Predicts Joint Effects of Aryl Hydrocarbon 
Receptor Agonists with Partial Agonists and Competitive Antagonists.  
Environmental Health Perspectives 118(5):666-672.\cr
Scholze, M. et al. 2001. A General Best-Fit Method for Concentration-Response Curves and the 
Estimation of Low-Effect Concentrations. Environmental Toxicology and Chemistry 
20(2):448-457.\cr
Wang, Yuan and Kai-Tai Fang. 1996. Uniform Design of Experiments with Mixtures. Science in China 
Series A-Mathematics Physics Astronomy 39(3):264-275.\cr
Backhaus, T., Faust, M., 2012. Predictive environmental risk assessment of chemical mixtures: A 
conceptual framework. Environmental Science and Technology. 46, 2564-2573.\cr
Zhu X-W, Liu S-S, Qin L-T, Chen F, Liu H-L. 2013. Modeling non-monotonic dose-response 
relationships: Model evaluation and hormetic quantities exploration. Ecotoxicology and 
Environmental Safety 89:130-136.\cr
Di Veroli GY, Fornari C, Goldlust I, Mills G, Koh SB, Bramhall JL, et al. 2015. An automated 
fitting procedure and software for dose-response curves with multiphasic features. 
Scitific Report 5: 14701.
}
\keyword{goodness of fit}
\keyword{curve fitting}
\keyword{concentration addition}
\keyword{independent action}
\keyword{generalized concentration addition}
